Move the CardContext into context folder. Update the paths.
